requirements.txt
redis==3.5.3
websocket-client==1.1.0
